buck2: log local / remote constraints for commands
Summary:
Like it says in the title. This would be useful to work out whether local
action caching is doing something for us (because it'll allow us to look at
whether actions that are local-only are now getting cache hits).

Note that this is actions-scale, but it's a very small amount of data.

Reviewed By: bobyangyf

Differential Revision: D39507846

fbshipit-source-id: 6b491a66dbcd553eb63620be9fe481a00b010864