Add `cxx_toolchain. archiver_supports_argfiles` to control archiver argfile usage
Summary:
Adds `archiver_supports_argfiles` field to `cxx_toolchain()` that can be used to control whether argfile are supported. This is required as some fbobjc libs started failing to archive on RE.

NOTE: BSD `ar` does not support argfile and other toolchains might not, either. So that's why we introduce a field to control the behavior explicitly.

Reviewed By: andrewjcg

Differential Revision: D39501574

fbshipit-source-id: 3720eb1522a7fbc713f8fa05b2e06d434033a59f