buck2: forkserver: expose cancellation
Summary:
Like it says in the title. This allows you to actually cancel things when
calling the forkserver. Actually using this is coming later.

Note have we could have done this when you drop your request, but the reason
we don't do this is because we need to know when the cancellation has taken
effect to avoid races.

Reviewed By: stepancheg

Differential Revision: D39547687

fbshipit-source-id: 72facf4aa7b0318fba51eaf55c2bf678c3114eca