buck2: forkserver: when terminating a process, kill the entire process group (on unix)
Summary:
This way, if subprocesses were spawned, we kill them too. This is not perfect
since subprocesses *could* have set their own process group, but while spawning
subprocesses is super common, setting a different process group isn't, so
that's a good step forward.

I meant to do this a while ago but it took a while to get `CommandExt::process_group`
landed so here we are finally :)

Reviewed By: stepancheg

Differential Revision: D39547690

fbshipit-source-id: 11dc0ea7247d58bbd64253c217e6d3e6ca0859b2