buck2: test: abort executions early if the test session is being shutdown
Summary:
Like it says in the title. There are 2 ways to do this:

- We can do what this does, which is to have specific points where we cancel
  (note: there are points where the cancellation *is* raced with execution,
  namely when we spawn a subprocess, that's done earlier in this stack).
- We can drop the futures.

Both have upsides and downsides:

- If we operate based on drop, we can abort anywhere. However, that means we
  need to specifically annotate places that *should not* be cancelled. Those
  are easy to forget (we already do this in some places with dropcancel
  "critical sections"). Since async drop is not a thing, that also means there
  can be no async cleanup, which is often a problem.
- If we make this explicit, then we can only abort in places where cancellation
  is checked.

I'm tempted to argue that explicit is better than implicit, and previous diffs
do always sometimes do drops when there is no risk, but I'm not entirely
certain yet. For now, this does the simple thing: put two checks at strategical
points (before materializations) and call it good enough.

Note that in practice we'll still kill things we're running if we didn't do
this, but it is better no to start them!

Reviewed By: stepancheg

Differential Revision: D39574195

fbshipit-source-id: 7388e796c442eaa37572f18e26484cff4208dbef