buck2: forkserver: make `Run` a streaming method
Summary:
I want to allow the caller to request cancellation. This is going to be used
for both tests and potentially hybrid execution. To do this, I need the caller
to be able to continue sending messages during execution, so that means making
Run a streaming method, which this does.

No behavior changes here, yet.

Reviewed By: stepancheg

Differential Revision: D39547692

fbshipit-source-id: b68635bda1247c9e8bfad6efe509c6a30bc5edca