buck2: attach a LivelinessManager to the CommandExecutionManager
Summary:
The goal of introducing this is to allow us to cancel commands when we don't
really care about them finishing anymore. This is, for example, true when
running tests after you CTRL+C out of the client. I also intend to experiment
with this for hybrid execution (by allowing the RE side to cancel local
commands).

To do this, this diff introduces a `LivelinessManager` which yields a future
that will not resolve while you are "alive", and will resolve once you're not.
We can then race that future with other things to decide when to cancel them
(but thats all coming later in this stack).

Reviewed By: stepancheg

Differential Revision: D39547691

fbshipit-source-id: 586c58860398a0a0b4f8f0ea2709706d89947d14