Add explicit branch for Xcode toolchain
Summary:
# Context

See D39539626 for context

# Changes

In this diff, we basically nest the Xcode toolchain constraints one level. This allows us

For correctness, we know that if any of the Xcode specific toolchain constraints were satisfied, the newly unversioned one would as well because it's defined as a subset of the buckconfig option (i.e., the unversion one requires just `"apple.toolchain": "apple-xcode-macos"` which is pre-requisite for the specific versioned ones).

Reviewed By: chatura-atapattu

Differential Revision: D39540632

fbshipit-source-id: 7df9b0ff2ba9877ba139b9c9e23133a981798b29