buck2: forkserver: explicitly close the output stream when we have a status
Summary:
Like it says in the title. Until now we were relying on the client going away
when they get the command result event, but it looks like this broke, and now
we keep this stream alive forever, which leads us to hold onto the underlying
FDs, and leads to leaking them:

https://fb.workplace.com/groups/buck2users/permalink/9595441457171050/

It's not super clear to me *why* this broke. I suspect it's due to D39547692 (https://github.com/facebookincubator/buck2/commit/670a979408260517785809c6dc1c76acaa7dca83),
which made this a streaming method. If I had to guess, I'd say that something
is probably keeping the response stream alive while the request stream is alive
(but sometimes it says alive forever). I'm basing that guess on the fact that
streams in Hyper must be Send + 'static, which suggests that they *are* being
sent to a different task.

Anyway, for now, this diff fixes the issue by doing the right thing and
explicitly reporting that our stream is done when we no longer have any input
to send.

Reviewed By: IanChilds

Differential Revision: D39653436

fbshipit-source-id: 182fc1b09f1e941091dcfab7ec0f313b5992fd40