Support comment markers on sections
Summary:
Currently Buck1 and Buck2 both only support full-line comments in .buckconfig files. That makes it hard to do OSS switching in these files, which we want to do.

I checked all the lines in the existing config and none have " #" as a substring, although plenty have "#" in them as a meaningful character. We could turn on comments everywhere, but that may have a perf impact, so instead just turn it on for sections in Buck2. That means we can use oss-enable in OSS, but not oss-disable, as that would break buck1.

Eventually, we should move to TOML, and not invent our own config format...

Reviewed By: stepancheg

Differential Revision: D39659504

fbshipit-source-id: 62c0e1f2e658c5ee912703902e2fac9b6949a932