Print RSS on IO superconsole page
Summary:
I wish we could display RSS on every page, but I could not find where to put it.

So place it on IO page. Following diff D39375530 adds CPU next to RSS.

This change is partly done to be able to quickly check it works correctly. I plan to show RSS in simpleconsole when buck stuck.

Reviewed By: bobyangyf

Differential Revision: D39371025

fbshipit-source-id: fb066495d06fee73e97779b442a46ade237af5bc