buck2: omnibus: sharing: add another common exclusion
Summary:
This is used by a lot of AI thins by virtue of being included as preload_deps
in a number of macros. This causes it to get excluded along with all its deps.

Excluding so much of Folly seems concerning. I think realistically we need
anonymous targets to make this palatable, but for now let's see if this can
unblock sore more contbuilds — we don't really have a better solution.

Reviewed By: ndmitchell

Differential Revision: D39698468

fbshipit-source-id: bf6758e8a2239886a2d9db1d270cff49632a7b8b