Fix examples/no-prelude/cpp (#13)
Summary:
The example included some old code that needed to be updated for compatibility with recent Starlark API changes.
At the moment the build does also have an implicit dependency on the prelude, so until that's fixed it's worked around by linking the prelude into the project.
Finally, the compiler toolchain introduced in the example was only used by the binary rule, but not by the library rule. This PR fixes that to use the toolchain consistently.

Pull Request resolved: https://github.com/facebookincubator/buck2/pull/13

Reviewed By: arlyon

Differential Revision: D39724852

Pulled By: arlyon

fbshipit-source-id: bd8ac4d3ce188dcb6b1f7cc98a09dad7c578242f