Migrate clean up logic from daemon side to client side
Summary:
`buck2 clean` currently does not work as expected on Windows. This happens because clean is executed by daemon and the same daemon process holds some files open. On Windows open files can't be deleted.

In order to get around this, we can just move all clean logic to the client side and kill the daemon before performing the logic.

Most of the logic here is just ported over from the server side clean implementation.

Reviewed By: krallin

Differential Revision: D39546367

fbshipit-source-id: 822f850726cfa42d0e6cb511e37b75e25f5bce28