merge_to_jar.py produces a deterministic output
Summary:
D39688961 (https://github.com/facebookincubator/buck2/commit/078a58f2b87773856d359b7af9850f9df0db51da) added an action so that we would scrub the timestamps of the things that merged into the jar.

That had a couple of issues:
(1) (minor) the output of this action still had the timestamps, so we could end up rerunning this action when we wouldn't need to otherwise.
(2) In some cases it looked like we can have a non-deterministic ordering when just giving the `jar` command a directory to add to an existing jar. Instead, give it a sorted list of paths to make that deterministic.

Reviewed By: milend

Differential Revision: D39727355

fbshipit-source-id: 0a7af4b2ffeff13af54d8800e1724e949f159658