bugfix to ocaml_object
Summary:
`ocaml_object` didn't have a test  that involved a non-empty `deps` list. the definition of the record `OCamlLibraryInfo` changed at some point in the past and `ocaml_object_impl` wasn't updated to account for that. the resulting code was then found later when exercised to be referencing field names on such records that didn't exist.

this diff fixes up the `ocaml_object_impl` to use proper field names and adds a test.

Differential Revision: D39744573

fbshipit-source-id: 255f42d1963a1fd85863060c258fad7a1a3604de