Make java_binary output deterministic
Summary:
The `jar` tool has an issue where the ordering of the contents of the jar is not guaranteed (https://bugs.openjdk.org/browse/JDK-8276764). This means that our `.jar` creation is non-deterministic and so we end up getting cache misses when we shouldn't (or don't need to, at least), because our tools "change".

We can fix the ordering by listing each of the files to be added to the jar individually. To do that we need to use an argfile as this command can get very big.

Reviewed By: jkeljo

Differential Revision: D39731167

fbshipit-source-id: e0c2d8485b94fd810dc944539d28c773e5a9e687