Make -a take an unanchored regex
Summary: We have `buck2 targets -a foo` and `buck2 cquery -a foo`. For targets, that means only return attributes which are precisely "foo". For cquery it means return attributes which match an unanchored regex "foo". The lack of consistency is deeply confusing. Given that moving to anchored for both would probably break a lot of stuff, instead move to unanchored for both, as I imagine a few more attributes won't break anything and it's easy enough to add "^foo$" if that's what you want.

Reviewed By: stepancheg

Differential Revision: D39807577

fbshipit-source-id: cdc141da83a7ad59ee0e02fb498fdce5d868d2da