Make order of mac exec platforms consistent
Summary:
NOTE: We kept the order in the previous diffs specifically because it made it easy to verify no change in behaviour as `uquery` output was identical.

Here, we adjust `_MAC_EXEC_PLATFORMS_INFO` to be consistent with `_MAC_EXEC_PLATFORM_KEYS`, so that we always list the minimal Xcode version first and then non-minimal.

Reviewed By: chatura-atapattu

Differential Revision: D39769992

fbshipit-source-id: 2d4eeec82c378254d417457e976debd9597974fc