buck2: hybrid executor: cancel local actions when remote is ready
Summary:
Like it says in the title. Using the hybrid executor as-is, we see some serious
slowdown when we try to do a large build (which doesn't get cache hits).

This happens because the local executor might steal an action then finish it
slower than RE (this can very easily happen if the action queued for a while),
and since we don't allow RE to take over if it finishes after local *starts*,
we wait for the (slow) local action.

To fix this, this diff updates the claim manager for RE to request cancellation
of local execution when it finishes. If local was executing, its command is
cancelled, it returns `ClaimCancelled`, and then we release the claim, and RE
can proceed.

Reviewed By: stepancheg

Differential Revision: D39658204

fbshipit-source-id: d0096acbab490e3ec916f61e05ea9aff0063f41c