buck2: hybrid executor: clean up some footguns
Summary:
Right now we construct the local_result before we know how it's going to be
used, which means that even when we're not actually racing the executors,
we have something constructed as if we were.

This means we need to do some awkward things, like explicitly dropping the
secondary in some places. It also means we have a bug: if you enable
the low pass filter and try to run an action that prefers local on a limited
hybrid executor then you'd deadlock.

To fix all this, let's rework the code so we put the filter last.

That would mean a lot of code duplication if we kept the structure of the code
as-is, since right now we were doing the "is it local preferred?" check once
but now we need it much later, so to mitigate this we put the futures into a
struct and add a function that does this for us.

This also neatly solves the "have to drop it" problem since that'll be
done implicitly by `into_primary()`.

Reviewed By: stepancheg

Differential Revision: D39851973

fbshipit-source-id: e9e8840019e3acaf4de62fa66e1e4ce2ffc41116