buck2: re-do claims implementation
Summary:
The current logic for claims is a little messy: the claim operation is atomic
and either succeeds or fails (it cannot blocks you), which means that if we
need to e.g. do a fallback, we have to restart the operation.

This makes it difficult to allow any kind of interaction between executors
because we have to restart the entire operation for them to interact. However,
I *do* want to have interactions to implement hybrid execution.

In particular, when the RE executor has output ready, I want it to be able to
ask the local executor to stop execution (if any was ongoing), and release its
claim. But, if local had finished executing, I want to do nothing, and if it
hadn't started, I want to prevent it from starting.

This is much easier to do if I can hook it up in the claim mechanism, since
then it means that my RE executor will kindly request control and will lock out
the local executor of execution (if it hadn't started), or otherwise wait until
the local executor returns, at which point we'll look at the result and either
unblock RE (if the local executor was indeed cancelled), or drop it (if the
local executor had finished).

So, this diff reworks our claims to have an API that will allow this. The next
diff uses it.

Reviewed By: ndmitchell

Differential Revision: D39658208

fbshipit-source-id: c5d7eafbd5d937dea10e569423f1ccbf4d804110