buck2: (poc) use the low-pass filter in hybrid execution
Summary:
This is all feature-gated, the intent is to start dogfooding this to see if
it produces desirable effects.

The intent of this particular implementation of hybrid is to optimize
for fast builds and minimal local resource usage by not using local
execution unless we know that given the low parallelism of the build
we are likely to win the race vs. RE.

"Low parallelism favors local" is a very basic heuristic, but if we
start to incorporate weights into this (and imagine a future where
we automatically guess those with some machine leanring!), it's
probably not that far off.

(I intend to POC a separate implementation that tries to minimize
duplicate work in order to optimize for resource usage across clients
and RE, I think the two should likely be separate things used in different
places)

Reviewed By: ndmitchell

Differential Revision: D39693924

fbshipit-source-id: bad4a8c15079738171bd73db0b9f318094ed1fd1