buck2: introduce "low-pass" filter for use in hybrid execution
Summary:
This diff introduces a weird kind of semaphore, which I'd like to land as a POC
to start dogfooding hybrid execution.

What this filter does is that it'll allow clients to enter as long as the
overall concurrency level is below a given number. If we exceed that number,
then the filter locks up for everyone (unlike a semaphore, which would still
allow a fixed number of clients through).

The idea is that when we only do local execution if there is a small
number of actions in flight. This is all very POC-grade (it's not aware
of e.g. the weight of actions).

Reviewed By: ndmitchell

Differential Revision: D39693925

fbshipit-source-id: 99b4883df58235fbd88936860e6ae54596dd380a