buck2: low pass filter: make the behavior of `access()` a bit more predictable
Summary:
When we use this, we use it in a `select()`, so there's a possibility it'll
never be polled (though the ordering of the futures in the select should ensure
that isn't the case).

This is a very specific detail to rely on so let's not do it and instead have
`access()` increment accessors when it's called and not when the future is
polled.

Reviewed By: ndmitchell

Differential Revision: D39771835

fbshipit-source-id: 3258b7fd59c85cc7c4404d78cdf877f38b55526b