buck2: ClaimManager: make `claim` take ownership + make release fallible
Summary:
Both of those API changes are necessary for the next diff.

With regard to the ownership, it intuitively makes sense: we know for sure we
can't claim the same thing twice, so consuming the manager is reasonable (but,
note, we might fail to obtain the claim even if we didn't claim it twice, if
the manager is internally mutable, which of course is the case here).

With regard to the fallibility, basically this ends up being necessary because
we sort of do some runtime lifetime checks here. The situation where can hit an
error is if we want to release RE's claim and we find out that local had
already started executing, which shouldn't be possible unless we withhold the
liveliness manager, which we don't do.

Reviewed By: stepancheg

Differential Revision: D39726191

fbshipit-source-id: 491fa24eb26a152e35bffb60a60f6ec54077f13d