buck2: forkserver: let the caller decide how their env var options combine
Summary:
Right now we have to duplicate the logic for priotizing clear, set and remove,
in the client side (for Windows) and the forkserver side (for Unix).

Instead, let's have it only on the client, and have the forkserver just receive
the order in which it should make calls. Concretely this makes it possible for
the caller to chose to either "Add X=Y, Remove X" or "Remove X, Add X=Y" and
have those things mean something meaningfully different (that is, in the
first case case you have no X and in the latter you do!).

Note that I'm not concerned with API compatibility here considering this is all
in the same binary anyway.

Reviewed By: lmvasquezg

Differential Revision: D39890148

fbshipit-source-id: cbfcec5b4669509bb80e8dad6090204a550c585f