Add more details about hidden resources in standalone python binaries
Summary:
Add a few more details to the error that is produced when a standalone python_binary depends on a non-python library with resources.

For build artifacts, show the source library.
For source artifacts, we don't have the rule that the file is bound to, so print its full path out so users know what files are causing issues.

Reviewed By: krallin

Differential Revision: D39828308

fbshipit-source-id: a281f2716f7a45ae5083040ff63e8f516f165b02