buck2: use a small string type for FileNameBuf
Summary:
Like it says in the title. This seems like it should be a good idea since in
isolation and other projects (e.g. hg, Watchman), it is known to perform better
so I'd like to just do it so don't have to revisit it later.

In practice, when building Adfinder, the data is a little bit all over the
place and I don't really discern an improvement or a regression.

Reviewed By: ndmitchell

Differential Revision: D39886331

fbshipit-source-id: 57cf405a46a6afac1ef1358a9757c3563ad54588