buck2: directory: use SmallMap instead of HashMap
Summary:
So that for small directories we don't allocate a hashmap.

Like for the previous diff, this seems like it should be a good idea since in
isolation and other projects (e.g. hg, Watchman), it is known to perform better
so I'd like to just do it so don't have to revisit it later.

In practice, when building Adfinder, the data is a little bit all over the
place and I don't really discern an improvement or a regression.

Reviewed By: ndmitchell

Differential Revision: D39886332

fbshipit-source-id: 8fb5d8684e6bdd94f86121c36ca6f9713ddcbb59