Include min sdk version in configuration we pass to deps of an Android binary
Summary:
When we do "pre-DEXing" on every individual jar in a binary, we need to pass the min SDK so that `d8` can do the right thing (in particular with `MethodHandle.invoke` and `MethodHandle.invokeExact` which are only supported on API level 26 and above).

In `buck2`, we usually pass information *up* the DAG. If we need to pass information *down* the graph, then we use configurations to accomplish that. Since we already have a transition at the binary level, we can just add a constraint value to the transition that will tell us which min SDK version we are building for.

Reviewed By: ianlevesque

Differential Revision: D39889457

fbshipit-source-id: 6e9381507988cd52b08f7cb2385ea07e0407fe5a