buck2: allow limiting the size of cache uploads
Summary:
Like it says in the title. Looking at our uploads I see we're sometimes
uploading things that probably don't make sense to cache (e.g. BTRFS images).

Realistically we probably want to annotate those as cacheable = False, but
it also wouldn't hurt to be able to add a sanity limit here. This will probably
be a necessity anyway to roll this out beyond just genrules.

Reviewed By: lmvasquezg

Differential Revision: D39927413

fbshipit-source-id: 2efa2fdbfe3d33c257ebb74a7cd2e6b03f4304a2