Parse nested_invocation buckconfig key
Summary:
I'm planning on adding two keys to fbsource and whatsapp buckconfigs:

```
[buck2]
  /// previous keys
  parallel_invocation = run | block | error
  nested_invocation = run | error
```

If these are not set, then the default is `run`. These keys only matter when the dice states are different:
* For `parallel_invocation`:
  * `run`: bypass the semaphore and run the commands concurrently without blocking
  * `block`: run the commands with blocking
  * `error`: panic
* For `nested_invocation`:
  * `run`: bypass the semaphore and run the commands concurrently without blocking
  * `error`: panic

When the dice state is the same, the commands will run concurrently for both nested invocations (with soft_error) and parallel invocations.

Next diff has parallel invocation handling.

Reviewed By: ndmitchell

Differential Revision: D39738110

fbshipit-source-id: 6f3d9cb2b4acbcc29da808fb58d2aba0ebc7d235