buck2: cxx: link: don't track external debug info if user explicitly asked for prefer_stripped_objects
Summary:
`prefer_stripped_objects` comes from `cpp_common` and it's set when `out_strip_mode == "full"`:
https://www.internalfb.com/code/fbsource/[4f2dd4abd78d65a44ca729fa2c53a9ae2dc3162e]/tools/build_defs/fbcode_macros/build_defs/lib/cpp_common.bzl?lines=1987-1988
(It's the only point in fbcode_macros that sets it).

If the user explicitly said they don't need debug info (`out_strip_mode == "full")` then we don't need to keep track of it and materialize outputs or enforce that they're packaged in `python_binary` resources.

 ---

Where I'm using it:

Packaging a link-group cxx_binary as a resource into a python_binary currently fails because the .o files are treated as runtime files (they have debug info) and aren't yet automatically packaged.

As a workaround we can strip the cxx_binary and package that, but configure_link_groups needs to also be stripped.

Propagate the binary's `strip_mode`.

Reviewed By: krallin

Differential Revision: D39912367

fbshipit-source-id: e6ac1b3426fe8fd3d32c7982593d64d4ad24f788