Don't warn when using cargo
Summary:
In the old days, some end users built Buck2 themselves with Cargo, and got a bad experience. Now we supply buck2 to users, anyone building it themselves must be a developer with a good sense of what they are doing, and even if not, they get what they expect. This warning is no longer helpful.

Also we don't want to show it for open source users, who should be using Cargo.

Reviewed By: stepancheg

Differential Revision: D39933050

fbshipit-source-id: cd45b4e5af13900bac9c5044b0df139b42c552d1