Explicitly do not support manifest_entries in buck2
Summary:
In `buck1`, `manifest_entries` is a grab-bag of things that were added to the manifest when running `aapt`. However, ever since we migrated to `aapt2`, they are no longer added to the manifest, so their name is quite misleading. Some aspects of the `manifest_entries` are being used elsewhere (e.g. the min SDK version) but it would be better to have an explicit attribute to make it clearer what's going on.

Therefore, going to explicitly not support `manifest_entries`, and add a `min_sdk_version` attribute in another diff.

Reviewed By: ianlevesque

Differential Revision: D39885821

fbshipit-source-id: e7ae4b649524f111552989e98901ecd841797f31