Make it explicit we ignore BuildSignalSender errors
Summary: Before we ignored them by taking the Result (which is must_use) and turning it to an option which we silently ignored. Now make it clear we are entirely ignoring the error. Would be ideal if we could explain why we were doing so, but at least it's now clear that we are doing so.

Reviewed By: IanChilds

Differential Revision: D39962874

fbshipit-source-id: 70149528e726177e2b18744149345a95fe1359fb