Unimplement StarlarkValue for str
Summary:
It prevents making `StarlarkValue` `Sized` which is needed to be able to add associated constants to `StarlarkValue`.

Associated constants are helpful because they can be accessed at compile time, unlike functions.

This can be helpful in changes like D36305002 for example, where short type name is placed in associated constant.

Also `str` is not really a `StarlarkValue` (it cannot be heap allocated).

Also with another implementation it is harder to verify that `str` implementation is consistent with `StarlarkStr` (because `StarlarkStr` caches the hash, and `str` does not).

If we need to use certain operations of `StarlarkStr` on `str`, we extract them into a separate functions.

Reviewed By: ndmitchell

Differential Revision: D36579300

fbshipit-source-id: 7f06c0a672d6d35d460729c8a6cf81be53230dd4