buck2: introduce a `CasArtifact` action
Summary:
Like it says in the title. This is an action that lets you reference a CAS artifact. Notionally it's a bit like download_file.

When the action executes it'll just verify that the content exists .

I made it so that you have to provide an minimum expiration timestamp when you add this to force users to think about the TTL of the artifacts they are referencing (though admittedly this was also an issue in download_file).

Reviewed By: stepancheg

Differential Revision: D39923155

fbshipit-source-id: 93588dca871eee582ca85176a125783a0977d559