Emit concurrent traces in ConcurrencyHandler
Summary:
Emitting all traces in the soft error message for easier debugging.

Also, I _think_ there's a deadlock if we error out for the `NestedInvocation::Error` case, and then try to run another command. The quickstack trace I saw all had the following:

```
Thread 215 (LWP 4048887):
#01  0x00007f3dabb259d9 in syscall () from /usr/local/fbcode/platform010/lib/libc.so.6
```
I'm not sure if this is related to something else, or it's because of the deadlock. Adding the lines to drop the data, create the drop guard, and then releasing the baton fixes it. I wasn't able to trigger this in my previous manual testing, but this seems like a correct thing to do regardless.

Reviewed By: krallin

Differential Revision: D39985753

fbshipit-source-id: 2e3bf9fa6ba5f340c0189aa9984c236fb80a6feb