spawn_background_process_on_windows
Summary:
Extract a function which spawn a background process on Windows.

Used in the following diff D40021507 but useful as is as an abstraction (for example, following diff D40024718 adds a test for it).

The function is available (but does not work) on Unix, because with this approach `cargo check` does more checks on Unix, making development faster without access to Windows machine. It is used in the following diff D40021507.

Reviewed By: KapJI

Differential Revision: D40021239

fbshipit-source-id: 76afac0d3055aa241ad7c84505e82377deb789e5