Move lock into drop guard in ConcurrencyHandler
Summary: In the previous diff, I was having issues with deadlock. Per Thomas's suggestion, it would be good to refactor the code so that future changes are more deadlock-resistant. We can achieve this by moving the lock into the drop guard so that we can make sure the drop guard can be instantiated with the lock, we don't need to `drop` the lock manually, and active traces are certain to be added and are able to be removed by the drop guard.

Reviewed By: krallin

Differential Revision: D40060969

fbshipit-source-id: 131e8a72d9b29767be89d5f767437c72e10d3a5f