Document exe_target
Summary: `exe_target` is how we run binaries without having to build them for the execution platform, and is often needed to prevent building binaries twice (once for the specified configuration of the binary rule, once for the execution platform when we run it). This adds a bit of documentation for `exe_target`.

Reviewed By: lmvasquezg

Differential Revision: D40104394

fbshipit-source-id: 9fc3e8f20796d920a380c9ccc21c232cc9aba055