Expose xcode version information in host_info
Summary:
For iOS builds, we need to be able to inform rules about the current xcode version (to pick appropriate toolchain versions).

In buck1, this is done by inspecting the state of /var/db/xcode_select_link, which points to the currently selected Xcode version (if any).

Buck2 can't arbitrarily introspect the state of the underlying host from rules, so we need to do this in buck2 core somewhere. One suggestion was to do this in `host_info()`.

This diff implements a new attribute - `xcode` on `host_info()`. In this attribute, we expose:
* Full Xcode version (major.minor.patch) as a string
* Major version as an int
* Minor version as an int
* Patch version as an int
* Whether the xcode is a "beta" version
* The Xcode build number

If we can't derive any of these values, we leave them set to `None`. In the case where no Xcode version is selected, _every_ attribute will be set to `None`.

Reviewed By: ndmitchell

Differential Revision: D39902385

fbshipit-source-id: b96e2e11b792ffc0f9c0f2691d5a982d2b464500