cfg(cargo_internal_build) instead of cfg(off)
Summary:
TIL, rustflags can be specified in `.cargo/config.toml`.

Define `--cfg cargo_internal_build` in that file.

Use it instead of `cfg(off)` for conditional compilation opensource vs internal.

It is better approach than `cfg(off)` because it is part of AST, so IDE and `rustfmt` handles it correctly.

It is better than features because it can be specified once globally.

But it is more verbose than `cfg(off)`.

Reviewed By: ndmitchell

Differential Revision: D40088750

fbshipit-source-id: 51a5ee2d8d11d7f48250752d78363b9203f2e625