Add a set of shims to compile open source
Summary:
We discussed the approach of shimming all the fbcode specific stuff, rewriting dependencies, and seeing how it went. I took that approach with Buck2 and was able to get Buck2 creating the action graph. It now fails for a few reasons (will outline them in a post), but this is a start towards making things work.

Eventually we'll want to clean up some pieces, especially the toolchains, but good enough for now.

Reviewed By: themarwhal

Differential Revision: D40002724

fbshipit-source-id: 4ad3dffc1e7c801fd81d724cd8be63044770b426