Constrain to bumpalo-3.8
Summary: Bumpalo-3.11 (the latest) requires a fixup to include the README as a source file, which is a bit annoying. Slightly surprisingly, if I require 3.8 then the Cargo.lock gets generated at 3.11, but I guess Cargo allows that range of movement in bounds. Use an explicit `=` to force it into 3.8, which doesn't require a fixup. Long term we'll want to move to a newer bumpalo, but no need to do that before everything else.

Reviewed By: stepancheg

Differential Revision: D40194128

fbshipit-source-id: faec6bb1231ed7edbf37e9f989dfc94879559c37