Convert _static_extension_utils from C++ to Cython
Summary: I was starting to get segmentation faults as I was working on converting some of the additional Python code to C. So I decided to look into converting the current C code into Cython, which actually simplifies it quite a bit and makes the rest of the migration trivial. It's also much easier to maintain the code, since most of them looks and behaves pretty much like Python.

Differential Revision: D40171356

fbshipit-source-id: 51348a6f3463f4e509a53da5694e567ad8c3928e