Add _build_only_native_code param to Java-like rules
Summary:
Per title, this adds a param that will be used in future diffs to prune the analysis we do when we only want to build native code.

These were previously added in the diffs where they are used, but that broke Intellij project generation - we need all of these to fix that in the next diff.

Reviewed By: ndmitchell

Differential Revision: D40260738

fbshipit-source-id: 08c4efe290d8c0c242106e77b88d9e095515c33b