Allocate dict with set capacity on attr coercion to reduce memory usage
Summary: When coercing selects, we allocate a new SmallMap for the coerced select but we don't set a fixed capacity for the SmallMap, even though we already know how big the map is ahead of time. This diff makes it allocates SmallMap to be same capacity as the original dict.

Reviewed By: stepancheg

Differential Revision: D40320396

fbshipit-source-id: 53a2e9cd87f4ebc65648ab96b4fc8ca76cc0ad5a