buck2: low_pass_filter: allow adding a weight to requests
Summary:
Like it says in the title. See the previous diff for a bit of context: overall,
it's good if we try to avoid sending more requests through the low_pass_filter
than the underlying host sharing broker would allow, so to do this we'll
put in the weights, because the host sharing broker uses them too.

The upshot is that if we have, say, 4 links that say their weight is 4, that'll
count for a concurrency of 16, not 4.

Reviewed By: stepancheg

Differential Revision: D40261016

fbshipit-source-id: 6f24887a44118d99af59fdcba925a9cf6118bff4