buck2: use the host concurrency for the low pass filter too
Summary:
It's a reasonable pick. Those don't have to be equal, but it doesn't make
much sense for the low pass filter to have a higher concurrency that the job
count. Rather than make the two independently configurable (which seems more
configuration than we need), let's just make them equal.

Reviewed By: scottcao

Differential Revision: D40261015

fbshipit-source-id: 73cbc8cef0f36af644e7db2271fab57b3c580bfa