Fix fat_jar and compile_and_package.py to produce deterministic ordering of files in their output jars
Summary:
D39976053 made it so that we add directories in our command, as well as files. For some reason, when adding a directory and then a file within that directory, we don't get a deterministic ordering.

We seem to get a deterministic ordering when adding all files and then all directories, so do that.

Reviewed By: navidqar

Differential Revision: D40340516

fbshipit-source-id: 38521b6a810e4d8ef98e34e6acee8c5bea05242d