optimize the opt workflow, copy entire folder instead of copy every single file when the archive has no bitcode object.
Summary:
Current opt step would check if each object in an archive is bitcode. then:
1. If it is, run an actual opt to generate ELF.
2. If it's not bitcode (already an ELF), trigger a copy action to copy the file to the destination.

In the case of prebuilt library since all the objects are ELF already, it's more efficient to just copy the folder than #2.

This change bring down the total number of actions from 186.5k -> 144k (22.7%) for unicorn:index_server

{F780561315}

Reviewed By: christycylee

Differential Revision: D40319823

fbshipit-source-id: 4650cf8f6ef5170d2c6fb4d456e7042de77a4f2f