Simplify static extension support, by using the native PyImport_Inittab and patching BuiltinImporter to support modules in packages
Summary:
So I ended up figuring this out by looking into `PyImport_ExtendInittab()` (which I had previosly looked into) and following the implementation to `_imp_create_module` which looked eerily familiar to code we had in `static_extension_utils.cpp`.

I had previously tried to add modules with packages to the `PyImport_Inittab` and noticed that this didn't work, but then I wondered why this was the case, since our implementation mostly matches that one in the `_imp` module and manually calling `find_spec` and `create_module` from `BuiltinImporter` seems to work to load modules within packages.

Then I noticed that this is what prevents that from working:

https://www.internalfb.com/code/fbsource/[e713bbf185765bf09c39f1730ecc681cbdc94c4b]/third-party/python/3.10/Lib/importlib/_bootstrap.py?lines=746-749

The check for `path is not None` ends up blocking modules within packages from working.

Tracking this code in history shows that it comes directly from [commit "Implement PEP 451 (ModuleSpec)"](https://github.com/python/cpython/commit/b523f8433a8982e10eb41a3e2b37ee0e6d6a6e00), so it seems this behavior has always been there.

Do we know why this is the case in the upstream `BuiltinImporter`? Is there a good reason to prevent builtin modules to be in packages? Perhaps some corner case we're missing, such as loading a module within a module (not package), or loading a module for an inexistent package?

In any case, it seems our loader/finder were already doing that, so it seems we're currently accepting changing this behavior.

So this finding, together with the fact that my `generate_static_extension_info.py` was already essentially generating the input to `PyImport_ExtendInittab()` (I just had to change the actual type used), made me think of simply reusing the native inittab support to implement static extensions.

That gets rid of most of our C++ code and usage of internal CPython APIs. Not to mention it simplifies this code significantly.

A few possible follow ups:

1. Investigate why upstream `BuiltinImporter` doesn't allow loading modules inside packages, and what it would take to possibly change this behavior upstream (removing the need for our custom `FullBuiltinImporter` with the patched `find_spec`.)

2. Looking into using a more efficient data structure for CPython's `PyImport_Inittab`, perhaps using a Python dictionary internally, just wrapping up the PyInit C functions into an appropriate Python object. This might become an issue if we start using the inittab for hundreds or thousands of builtin modules. There's also a list of builtin modules exported in `sys`, which could be turned into a Python set for faster lookup. This change can also be done upstream.

#forcetdhashing

Reviewed By: thatch

Differential Revision: D40233238

fbshipit-source-id: efa03daaa7979b17accc85838206f201b346bf2e